requirements.txt
alabaster==0.7.12
Babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coverage==5.0.4
cryptography==2.9
docutils==0.16
idna==2.9
imagesize==1.2.0
Jinja2==2.11.1
MarkupSafe==1.1.1
packaging==20.3
pip==20.0.2
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-futures==1.0.0
requests-mock==1.7.0
requests-toolbelt==0.9.1
-e git+https://github.com/WIPACrepo/rest-tools@cdfeacdaafd546ed96821ce57d646146ed75a9e4#egg=rest_tools
setuptools==46.1.3
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tornado==6.0.4
urllib3==1.25.8
